+++ src/index.ts
@@ -285,6 +285,8 @@ export {
 } from "./components/skeleton/index.js";
 export * from "./components/progress/index.js";
 export * from "./components/progress-circle/index.js";
+export * from "./components/slider/index.js";
+export * from "./components/rating/index.js";
 export * from "./components/toast/index.js";
 export * from "./components/collapsible/index.js";
 export * from "./components/accordion/index.js";